Updated the beamline vaccume of the lh2 20cm, 4cm and C12 setups. I had forgotten to fill the vaccume inside the beampipe after seperating the beampipe into 2 sections. The 2 sections was required to fit the new steel pipe and collar Michael Moor is proposing as a fix to the fringe field issue. Anyway, I forgot to fill the second half of the beampipe (from bender to He dump vessel) with vaccuum. This commit fix that issue.
The new setup lh2_20cm_with_hb_shield.gdml contains a 10 cm thick tungsten shield placed infront of the bender (right on the front face of the cryostat). This setup is being used to see if the shield can reduce the heat deposition on the beam side of the bender coils.